Ask for delete confirmation for attachment. Delete folder for attachments when all attachments have been removed.

+++ Modules/datacrow-client/src/main/java/org/datacrow/client/console/components/panels/AttachmentsPanel.java
@@ -110,10 +110,14 @@ public class AttachmentsPanel extends DcPanel implements MouseListener, ActionLi
 
     private void deleteAttachments() {
     	List<Attachment> attachments = list.getSelectedAttachments();
-    	for (Attachment attachment : attachments)
-    		deleteAttachment(attachment);
-    	
-    	list.clearSelection();
+    	String msg = attachments.size() > 1 ? "msgDeleteAttachmentsConfirmation" : "msgDeleteAttachmentConfirmation";
+    	   	
+    	if (GUI.getInstance().displayQuestion(msg)) {
+	    	for (Attachment attachment : attachments)
+	    		deleteAttachment(attachment);
+	    	
+	    	list.clearSelection();
+    	}
     }
     
     private void deleteAttachment(Attachment attachment) {
